zsh: Fix post-up hook
Annoyingly, the zsh shell only supports the 'dotted' names .zprofile and
.zshrc even when they are placed in the $XDG_CONFIG_DIR (ie. ~/.config).
However, it is far more convenient to edit these files without the
'dotted' prefix (as the dot-prefix indicates to the view that the file
should be hidden by default).

To handle this better, the true files are the 'undotted' versions, and
the post-up hook simply adds symbolic links to mate them up with the
'dotted' names that zsh expects. Editing through either name results in
accessing the same content.